Move "--protected-mode no" into the entrypoint instead
Since doubling up on this option will have the "latest" version win, we're safe (in the sense of an image user being able to still override it) to apply this to all invocations of the server.

This change also includes the glue necessary to allow for `docker run redis --flag`, which frustrated me while I was testing this change.