Change the redis group ID to 998 on alpine
The GID 1000 is normally assigned as the default for the first user
created on linux distributions. Having it already in the image makes
it so injecting the host user and group into a custom image using redis
as base fail, requiring workarounds.

Having it use a GID below 1000 is preferable.