Update templating to use a single template
This also converts us to just two variants and auto-detects the latest Alpine/Debian, since we don't have a good reason to support more than one of each Debian and Alpine.

Additionally, this auto-scrapes the list of checksums and supported architectures for `gosu` from the upstream project (since if we're going to embed the version in `versions.json`, we might as well go all the way and pre-scrape whatever data we possibly can), and switches to use `gosu` consistently between variants (given the state of `su-exec`).